requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
PyYAML==6.0.1
replicate==0.23.1
Requests==2.31.0
